create services and products.routes

+++ packages/server/src/app/controllers/DepartmentsController.js
@@ -1,75 +1,54 @@
 import Department from "../schemas/Department";
-import AppError from '../shared/errors/AppError';
+import Product from "../schemas/Product";
+import AppError from "../shared/errors/AppError";
 
 class DepartmentController {
   async index(_, res) {
-    try {
-      const departments = await Department.find();
-      return res.json(departments);
-    } catch (err) {
-      return res.status(500).json({
-        message: {
-          errorMessage: "Internal Server Error",
-        },
-      });
-    }
+    const departments = await Department.find();
+    return res.json(departments);
   }
 
   async store(req, res) {
-    try {
-      const { name } = req.body;
-      const departmentSchema = new Department({ name });
-      const department = await departmentSchema.save();
-      return res.json(department);
-    } catch (err) {
-      return res.status(500).json({
-        message: {
-          errorMessage: "Internal Server Error",
-        },
-      });
-    }
+    const { name } = req.body;
+    const departmentSchema = new Department({ name });
+    const department = await departmentSchema.save();
+    return res.json(department);
   }
 
   async update(req, res) {
-    try {
-      const { id } = req.params;
-      const department = await Department.findById({ _id: id });
+    const { id } = req.params;
+    const department = await Department.findById({ _id: id });
 
-      if (!department) {
-        return res.status(400).json({
-          message: {
-            errorMessage: "Not found",
-          },
-        });
-      }
-
-      department.name = req.body.name;
-      await department.save();
-
-      return res.json(department);
-      
-    } catch (err) {
-      console.log(err);
-      return res.status(500).json({
-        message: {
-          errorMessage: "Internal Server Error",
-        },
-      });
+    if (!department) {
+      throw new AppError("Department not found");
     }
+
+    department.name = req.body.name;
+    await department.save();
+
+    return res.json(department);
   }
 
   async delete(req, res) {
-    try {
-      const { id } = req.params;
-      await Department.deleteOne({ _id: id });
-      return res.status(204).send("");
-    } catch (err) {
-      return res.status(500).json({
-        error: {
-          message: "Internal Server Error",
-        },
-      });
+    const { id } = req.params;
+
+    const department = await Department.findById({ _id: id });
+
+    if (!department) {
+      throw new AppError("Department not found");
     }
+
+    const products = await Product.find({ departments: id }).exec();
+
+    if (products.length > 0) {
+      throw new AppError(
+        "Could not remove this department. You may have to fix its dependencies before."
+      );
+    }
+
+    await Department.deleteOne({ _id: id });
+
+    return res.status(204).send();
   }
 }
